split proto tests
To simple, xproto, copy
For better CI job results visibility
