+++ test/xproto/Dockerfile
@@ -3,4 +3,4 @@ FROM golang:latest
 WORKDIR /root
 COPY . /root
 
-ENTRYPOINT [ "go", "test", "./test/xproto" ]
+ENTRYPOINT ["sh", "-c", "go test -tags=${TEST_TAG:-all} -v ./test/xproto/..."]
